Update UpdateContacts.php
Fixed issue where if the user did not make any changes to the contact it would respond with "No records found". This is because the an update query can occur even if the contact does not exist and the previous solution to that was to check if there were any rows affected, but if a user did not make a change to the contact it would respond with "No records found", because no rows were affected. New solution is to run a SELECT query for the desired contact to update and then check to see if it exists, if not then return "No records found".